色々修正

+++ src/main/java/red/man10/man10itembank/ItemData.kt
@@ -3,8 +3,6 @@ package red.man10.man10itembank
 import org.bukkit.Bukkit
 import org.bukkit.entity.Player
 import org.bukkit.inventory.ItemStack
-import red.man10.man10itembank.ItemData.CallBack
-import red.man10.man10itembank.ItemData.Transaction
 import red.man10.man10itembank.util.MySQLManager
 import red.man10.man10itembank.util.MySQLManager.Companion.escapeStringForMySQL
 import red.man10.man10itembank.util.Utility
@@ -16,11 +14,23 @@ import java.util.concurrent.LinkedBlockingQueue
 
 object ItemData {
 
-    private val transactionQueue  = LinkedBlockingQueue<Pair<Transaction,CallBack>>()
+    private val transactionQueue  = LinkedBlockingQueue<()->Unit>()
     private var itemIndex = ConcurrentHashMap<Int,ItemIndex>()
+
+    private var transactionThread = Thread{ transaction() }
 
     private lateinit var mysql : MySQLManager
 
+    init {
+        runTransactionQueue()
+
+        val transaction = Transaction@{
+            asyncLoadItemIndex()
+        }
+
+        addTransaction(transaction)
+
+    }
     //キューが詰まった時に確認するためのもの
     fun getQueueSize():Int{
         return transactionQueue.size
@@ -33,40 +43,30 @@ object ItemData {
     //ItemStackからItemIndex取得
     fun getItemData(item: ItemStack):ItemIndex?{
 
-        var data : ItemIndex? = null
-
-        itemIndex.forEach{
-            if (item.isSimilar(it.value.item)){
-                data = it.value
-            }
-        }
-        return data
+        return itemIndex.filter { it.value.item?.isSimilar(item)?:false }.getOrDefault(0, null)
     }
 
     //KeyからItemIndex取得
-    fun getItemData(key:String):ItemIndex?{
+    fun getItemData(key: String): ItemIndex? {
+        return itemIndex.filter { it.value.itemKey == key }.getOrDefault(0, null)
+    }
 
-        var data : ItemIndex? = null
-
-        itemIndex.forEach{
-            if (it.value.itemKey==key){
-                data = it.value
-            }
-        }
-        return data
+    fun getID(key: String):Int{
+        return getItemData(key)?.id?:-1
     }
 
     fun getItemIndexMap():Map<Int,ItemIndex>{
         return itemIndex
     }
 
-    //ItemIndexに新規アイテムを登録 0:成功、1:重複、2:失敗
-    fun registerItem(player: Player, key:String, item:ItemStack,initialPrice:Double,tick:Double,callBack: CallBack = CallBack {}){
+    //ItemIndexに新規アイテムを登録
+    fun registerItem(player: Player, key:String, item:ItemStack,initialPrice:Double,tick:Double,callBack: (EnumResult)->Unit = {}){
 
-        val transaction = Transaction {
+        val transaction = Transaction@{
 
             if (getItemData(key)!=null){
-                return@Transaction 1
+                callBack.invoke(EnumResult.FAILED)
+                return@Transaction
             }
 
             mysql.execute("INSERT INTO item_index (item_key, item_name, price, bid, ask, tick, time, disabled, base64) " +
@@ -76,21 +76,29 @@ object ItemData {
 
             if (getItemData(key)!=null){
                 log("アイテムインデックス登録 item_key:${key}",player)
-                return@Transaction 0
+                callBack.invoke(EnumResult.SUCCESSFUL)
+                return@Transaction
             }
 
-            return@Transaction 2
+            callBack.invoke(EnumResult.FAILED)
+            return@Transaction
         }
 
-        addTransaction(transaction,callBack)
+        addTransaction(transaction)
     }
 
     //アイテム削除 0:成功、1:存在しない、2:失敗
-    fun unregisterItem(player: Player,id:Int,callBack: CallBack = CallBack {}){
+    fun unregisterItem(player: Player,id:Int,callBack: (EnumResult)->Unit = {}){
 
-        val transaction = Transaction{
+        val transaction = Transaction@{
 
-            val data = getItemData(id)?:return@Transaction 1
+            val data = getItemData(id)
+
+            if (data==null){
+                callBack.invoke(EnumResult.FAILED)
+                return@Transaction
+            }
+
             val key = data.itemKey
 
             mysql.execute("DELETE FROM item_index WHERE id=${id};")
@@ -98,24 +106,27 @@ object ItemData {
 
             if (getItemData(id) == null){
                 log("アイテムインデックス削除 item_key:${key}",player)
-                return@Transaction 0
+                callBack.invoke(EnumResult.SUCCESSFUL)
+                return@Transaction
             }
 
-            return@Transaction 2
+            callBack.invoke(EnumResult.FAILED)
+            return@Transaction
         }
 
-        addTransaction(transaction, callBack)
+        addTransaction(transaction)
 
     }
 
-    //返り値は在庫 -1は失敗or存在しないID
-    fun addItemAmount(order:UUID?,target:UUID, id: Int, amount: Int, callBack: CallBack = CallBack {}){
+    //返り値は在庫 nullは失敗
+    fun addItemAmount(order:UUID?,target:UUID, id: Int, amount: Int, callBack : (Int?)->Unit = {}){
 
-        val transaction = Transaction {
+        val transaction = Transaction@ {
             val nowAmount = asyncGetItemAmount(target, id)
 
-            if (nowAmount == -1){
-                return@Transaction -1
+            if (nowAmount == null){
+                callBack.invoke(null)
+                return@Transaction
             }
 
             val newAmount = nowAmount+amount
@@ -124,88 +135,95 @@ object ItemData {
 
             Log.storageLog(order,target,id,amount,newAmount,"AddItem")
 
-            return@Transaction newAmount
+            callBack.invoke(newAmount)
+            return@Transaction
         }
 
-        addTransaction(transaction, callBack)
+        addTransaction(transaction)
 
     }
 
-    //返り値は在庫(取り出せなかった場合は-2)
-    fun takeItemAmount(order: UUID?, target:UUID, id: Int, amount: Int, callBack: CallBack = CallBack {}){
+    //返り値は在庫(取り出せなかった場合は-1)
+    fun takeItemAmount(order: UUID?, target:UUID, id: Int, amount: Int, callBack: (Int?)->Unit =  {}){
 
-        val transaction = Transaction {
+        val transaction = Transaction@ {
             val nowAmount = asyncGetItemAmount(target, id)
 
-            if (nowAmount == -1){
-                return@Transaction -1
+            if (nowAmount == null){
+                callBack.invoke(null)
+                return@Transaction
             }
 
             val newAmount = nowAmount-amount
 
             //取り出し失敗
             if (newAmount < 0){
-                return@Transaction -2
+                callBack.invoke(-1)
+                return@Transaction
             }
             asyncSetItemAmount(target, id, newAmount)
             Log.storageLog(order,target,id,amount,newAmount,"TakeItem")
 
-            return@Transaction newAmount
+            callBack.invoke(newAmount)
+
+            return@Transaction
         }
 
-        addTransaction(transaction, callBack)
+        addTransaction(transaction)
 
     }
 
     //アイテム数を設定する 返り値は最新の在庫
-    fun setItemAmount(order: UUID?,target:UUID, id: Int, amount: Int, callBack: CallBack = CallBack {}){
+    fun setItemAmount(order: UUID?,target:UUID, id: Int, amount: Int, callBack: (Int?)->Unit = {}){
 
         var newAmount = amount
 
         if (amount <0){ newAmount = 0 }
 
-        val transaction = Transaction {
+        val transaction = Transaction@ {
 
             //倉庫の有無をチェック
             val nowAmount = asyncGetItemAmount(target, id)
 
             if (nowAmount == -1){
-                return@Transaction -1
+                callBack.invoke(null)
+                return@Transaction
             }
 
             asyncSetItemAmount(target, id, newAmount)
             Log.storageLog(order,target,id,amount,newAmount,"SetItem")
 
-            return@Transaction amount
+            callBack.invoke(amount)
+            return@Transaction
         }
 
-        addTransaction(transaction,callBack)
+        addTransaction(transaction)
 
     }
 
     //アイテム数を取得する、callBack関数は、取得後に走る処理、引数のresultが取得した数
-    fun getItemAmount(uuid:UUID,id: Int,callBack: CallBack = CallBack {}){
-        val transaction = Transaction {
-            return@Transaction asyncGetItemAmount(uuid, id)
+    fun getItemAmount(uuid:UUID,id: Int,callBack: (Int)->Unit = {}){
+        val transaction = Transaction@ {
+            callBack.invoke(asyncGetItemAmount(uuid, id))
         }
-        addTransaction(transaction, callBack)
+        addTransaction(transaction)
     }
 
 /////////////////////////////////キューの中で、キューに突っ込む処理を入れないこと(キューが詰まるため)////////////////////////////////////////
 
     //同時に実行されてほしくない処理は、ここに処理を投げる
-    private fun addTransaction(transaction:Transaction,callBack:CallBack= CallBack {}){
-        transactionQueue.add(Pair(transaction,callBack))
+    private fun addTransaction(transaction:()->Unit){
+        transactionQueue.add(transaction)
     }
 
     //在庫を設定
-    private fun asyncSetItemAmount(uuid: UUID,id: Int,amount:Int):Int{
+    private fun asyncSetItemAmount(uuid: UUID,id: Int,amount:Int):Boolean{
 
-        getItemData(id) ?: return -1
+        getItemData(id) ?: return false
 
         mysql.execute("UPDATE item_storage SET amount = $amount,time=now() WHERE uuid='${uuid}' and item_id=${id};")
 
-        return 0
+        return true
     }
 
     //在庫を取得
@@ -218,7 +236,7 @@ object ItemData {
         //倉庫が存在しなかった場合
         if (!rs.next()){
             asyncCreateItemStorage(uuid,id)
-            return asyncGetItemAmount(uuid, id)
+            return 0
         }
 
         val amount = rs.getInt("amount")
@@ -230,9 +248,9 @@ object ItemData {
     }
 
     //アイテムバンクを作成 0:成功 2:失敗
-    private fun asyncCreateItemStorage(uuid: UUID,id:Int):Int{
+    private fun asyncCreateItemStorage(uuid: UUID,id:Int):EnumResult{
 
-        val data = getItemData(id) ?: return 2
+        val data = getItemData(id) ?: return EnumResult.FAILED
 
         val p = Bukkit.getOfflinePlayer(uuid)
 
@@ -241,10 +259,10 @@ object ItemData {
 
         if (asyncGetItemAmount(uuid,id) != -1) {
             Log.storageLog(null,uuid,id,0,0,"CreateStorage")
-            return 0
+            return EnumResult.SUCCESSFUL
         }
 
-        return 2
+        return EnumResult.FAILED
     }
 
     //ItemIndexの読み込み
@@ -286,34 +304,43 @@ object ItemData {
     }
 
 
-    fun interface Transaction{ fun onTransactionResult():Int}
-    fun interface CallBack{ fun callBack(result: Int) }
-
     private fun runTransactionQueue(){
+
+        interruptTransactionQueue()
+
+        transactionThread = Thread{ transaction() }
+        transactionThread.start()
+
+    }
+
+    private fun interruptTransactionQueue(){
+
+        if (transactionThread.isAlive)
+            transactionThread.interrupt()
+
+    }
+
+    private fun transaction(){
 
         log("トランザクションキュー起動")
 
-        Thread{
+        mysql = MySQLManager(Man10ItemBank.plugin,"Man10ItemBankTransaction")
 
-            mysql = MySQLManager(Man10ItemBank.plugin,"Man10ItemBankTransaction")
+        while (true){
 
-            while (true){
+            try {
+                val queue = transactionQueue.take()
 
-                try {
-                    val queue = transactionQueue.take()
+                queue.invoke()
 
-                    val result = queue.first.onTransactionResult()
+            }catch (e:InterruptedException){
+              log("トランザクションの停止")
+              break
+            } catch (e:Exception){
+                log("トランザクションキューのエラー:${e.stackTrace}")
 
-                    queue.second.callBack(result)
-
-                }catch (e:Exception){
-                    log("トランザクションキューのエラー:${e.stackTrace}")
-
-                }
             }
-
-        }.start()
-
+        }
     }
 
     class ItemIndex{
@@ -327,16 +354,12 @@ object ItemData {
         var item : ItemStack? = null
     }
 
-    init {
-        runTransactionQueue()
-
-        val transaction = Transaction{
-            asyncLoadItemIndex()
-            return@Transaction 0
-        }
-
-        addTransaction(transaction)
+    enum class EnumResult{
+        SUCCESSFUL,
+        FAILED,
 
     }
+
+
 
 }
